+++ lib/omnisharp-atom/services/apply-changes.ts
