+++ lib/omnisharp-atom/services/completion-provider.ts
