+++ lib/omnisharp-atom/services/framework-provider.ts
