+++ lib/omnisharp-atom/services/linter-provider.ts
