+++ lib/omnisharp-atom/services/omni-select-list-view.ts
