+++ lib/omnisharp-atom/services/project-provider.ts
