+++ lib/omnisharp-atom/views/find-symbols-view.ts
@@ -1,4 +1,4 @@
-import OmniSelectListView = require('../providers/omni-select-list-view');
+import OmniSelectListView = require('../services/omni-select-list-view');
 import Omni = require('../../omni-sharp-server/omni');
 
 class FindSymbolsView extends OmniSelectListView {
